Upgrade project to stb 0.13
