+++ project/BuildSettings.scala
@@ -1,0 +1,25 @@
+import sbt.Keys._
+import sbt._
+import sbtassembly.AssemblyPlugin.autoImport._
+
+object BuildSettings {
+  val buildOrganization = "cfpb"
+  val buildVersion      = "1.0.0"
+  val buildScalaVersion = "2.11.8"
+
+  val hmdaBuildSettings = Defaults.coreDefaultSettings ++
+    Seq(
+      organization := buildOrganization,
+      version      := buildVersion,
+      scalaVersion := buildScalaVersion,
+      scalacOptions ++= Seq(
+        "-Xlint",
+        "-deprecation",
+        "-unchecked",
+        "-feature"),
+      aggregate in assembly := false,
+      parallelExecution in Test := false,
+      testOptions in Test += Tests.Argument(TestFrameworks.ScalaTest, "-oC")
+    )
+
+}
